remove newsletter modal component, we now use a separate page instead
