--- components/modal/newsletter-modal/index.js
@@ -1,3 +1,0 @@
-import NewsletterModalComponent from './component';
-
-export default NewsletterModalComponent
