--- css/components/modal/newsletter_modal.scss
@@ -1,5 +1,0 @@
-.c-newsletter-modal {
-  iframe {
-    border: 0;
-  }
-}
