+++ pages/app/newsletter/component.js
@@ -15,7 +15,7 @@ import Select from 'components/form/SelectInput';
 import { Link } from 'routes';
 
 // Constants
-import { FORM_COUNTRIES } from 'components/modal/newsletter-modal/constants';
+import { FORM_COUNTRIES } from 'pages/app/newsletter/constants';
 
 class Newsletter extends Page {
   render() {
